datus-db-adapters/datus-clickzetta/pyproject.toml
[project]
name = "datus-clickzetta"
version = "0.1.0"
description = "ClickZetta database adapter for Datus"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "DatusAI", email = "support@datus.ai"}
]
keywords = ["datus", "database", "clickzetta", "adapter", "lakehouse"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "datus-agent>=0.2.2",
    "clickzetta-connector-python",
    "clickzetta-zettapark-python",
    "pandas>=2.0.0",
    "pyarrow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/datusai/Datus-adapters"
Repository = "https://github.com/datusai/Datus-adapters"
Issues = "https://github.com/datusai/Datus-adapters/issues"

[project.entry-points."datus.adapters"]
clickzetta = "datus_clickzetta:register"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["datus_clickzetta"]